# back/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Canadian Solar ERP API",
    "description": "API complète pour la gestion de magasin solaire et électronique",
    "version": "1.0.0",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Serveur de développement"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "CashTransaction": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "type": "string",
            "enum": [
              "income",
              "expense"
            ]
          },
          "amount": {
            "type": "number",
            "format": "decimal"
          },
          "description": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "customerId": {
            "type": "string",
            "format": "uuid"
          },
          "customerName": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/finance/transactions": {
      "get": {
        "tags": [
          "Finance"
        ],
        "summary": "Lister les transactions financières",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des transactions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CashTransaction"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Finance"
        ],
        "summary": "Créer une transaction",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CashTransaction"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Transaction créée"
          },
          "400": {
            "description": "Données invalides"
          }
        }
      }
    },
    "/api/finance/daily-report": {
      "get": {
        "tags": [
          "Finance"
        ],
        "summary": "Rapport journalier",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "startDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Rapport journalier"
          }
        }
      }
    },
...
    "/api/finance/transactions/stats": {
      "get": {
        "tags": [
          "Finance"
        ],
        "summary": "Statistiques financières",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "startDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Statistiques"
          }
        }
      }
    },
    "/api/finance/transactions/{id}": {
      "put": {
        "tags": [
          "Finance"
        ],
        "summary": "Mettre à jour une transaction",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Transaction mise à jour"
          }
        }
      },
      "delete": {
        "tags": [
          "Finance"
        ],
        "summary": "Supprimer une transaction",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Transaction supprimée"
          }
        }
      }
    },
    "/api/finance/transactions/{id}/comment": {
      "put": {
        "tags": [
          "Finance"
        ],
        "summary": "Commenter une transaction",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "comment": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Commentaire ajouté"
          }
        }
      }
    },
    "/api/finance/flux-journalier": {
      "get": {
        "tags": [
          "Finance"
        ],
        "summary": "Flux journalier",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "startDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Flux journalier"
          }
        }
      }
    },
...
  }
}